compile two different css files
Compiles the original phenotypes css file which is usable in any browser and is non-themable. Also compiles a separate second phenotypes css file which contains the css variables and is not compatible with any browser that doesn't support css variables. In order to make it work with all browsers a ponyfill must be used.